+++ core/src/main/java/com/crowdar/driver/DriverManager.java
@@ -20,18 +20,12 @@ public class DriverManager {
 
     
     public static void initialize(Map<String, ?> extraCapabilities){
-//    	String created = String.valueOf(isDriverCreated());
-    	
     	if (!isDriverCreated() || !isAValidDriver()) {
             if (localDriver.get() != null) {
                 localDriver.remove();
             }
             localDriver.set(DriverFactory.createDriver(extraCapabilities));
         }
-    	
-//        String logTemplate = "######  %s - Thread id %s --- isDriverCreated %s --- DriverId %s";
-//        System.out.println(logTemplate.format(logTemplate, "Initialize", Thread.currentThread().getId(), created, localDriver.get().getSessionId()));
-        
     }
     
     public static void initialize(ProjectTypeEnum projectType, SetupStrategy setupStrategy, URL driverHub, Map<String, ?> extraCapabilities) throws Exception {
@@ -47,26 +41,16 @@ public class DriverManager {
     }
 
     public static RemoteWebDriver getDriverInstance() {
-    	
-//    	String created = String.valueOf(isDriverCreated());
-    	
         if (!isDriverCreated() || !isAValidDriver()) {
             if (localDriver.get() != null) {
                 localDriver.remove();
             }
             localDriver.set(DriverFactory.createDriver());
         }
-        
-//        String logTemplate = "######  %s - Thread id %s --- isDriverCreated %s --- DriverId %s";
-//        System.out.println(logTemplate.format(logTemplate, "GetInstance", Thread.currentThread().getId(), created, localDriver.get().getSessionId()));
-        
         return localDriver.get();
     }
 
     public static void dismissCurrentDriver() {
-//    	String logTemplate = "######  %s - Thread id %s --- isDriverCreated %s --- DriverId %s";
-//    	System.out.println(logTemplate.format(logTemplate, "DismissDriver", Thread.currentThread().getId(), isDriverCreated(), localDriver.get().getSessionId()));
-
     	if (isDriverCreated()) {
             localDriver.get().quit();
             localDriver.remove();